# disable gost key mask on gost94 digest steps due to performance issues