@vt-alt 2020-09-23
Making the key is not needed for digest - that sounds reasonable.